Improve support for windows containers on copy to and from windows containers
Docker versions before version 1.8 did not support docker cp for copying files, so "dd" was used
to copy files to and from the container. Windows does not have "dd" so that failed in windows containers.

Now we docker inspect to determine the container platform, and when the platform is "windows", we try to
copy the files with docker cp - and else we revert to the previous behaviour.

With this change we can also utilize windows containers for testing with molecule with the docker driver which uses
this community.docker collection to manage containers.